TcpServer,Acceptor,TcpConnection done Connector ing

+++ src/net/SockAddrIn.cc
@@ -6,5 +6,5 @@ SockAddrIn::SockAddrIn(int port) {
     bzero(&addr_, sizeof(addr_));
     addr_.sin_family = AF_INET;
     addr_.sin_addr.s_addr = htonl(INADDR_ANY);
-    addr_.sin_port = htons(port);
+    addr_.sin_port = htons(static_cast<uint16_t>(port));
 }
